fixed the code for auth
