+++ src/services/api/auth.ts
@@ -1,11 +1,24 @@
-import { axiosInstance } from ".";
-import { AdminLogin } from "../../types/type";
+import { AxiosResponse } from "axios";
 
-export const login = async (data: AdminLogin) => {
+import { axiosInstance } from ".";
+import { ILoginResponse } from "../../types/authTypes";
+
+export const handleLogin = async ({
+  email,
+  password,
+}: {
+  email: string;
+  password: string;
+}) => {
   try {
-    return await axiosInstance.post("auth/signIn", data);
+    const response: AxiosResponse<ILoginResponse> =
+      await axiosInstance.post<ILoginResponse>(`auth/signIn`, {
+        email,
+        password,
+      });
+    return response.data;
   } catch (error) {
-    console.error(error);
+    console.error("Error in handleLogin function:", error);
     throw error;
   }
 };
