+++ src/services/mutations/authMutation.ts
@@ -1,19 +1,30 @@
 import { useMutation } from "@tanstack/react-query";
-import { login } from "../api/auth";
-import { AdminLogin } from "../../types/type";
-import { useAuthStore } from "../../store/authStore";
+import axios from "axios";
+import { toast } from "sonner";
+import { handleLogin } from "../api/auth";
 
-export function useLogin() {
-  const setToken = useAuthStore((state) => state.setToken);
-
+export function useLoginUser() {
   return useMutation({
-    mutationFn: (data: AdminLogin) => login(data),
-    onSuccess: (response) => {
-      const { access_token, refresh_token } = response.data.data.token;
-      setToken(access_token);
-      localStorage.setItem("refresh_token", refresh_token);
-      console.log("Access Token:", access_token);
-      console.log("Refresh Token:", refresh_token);
+    mutationFn: async ({
+      email,
+      password,
+    }: {
+      email: string;
+      password: string;
+    }) => {
+      return handleLogin({ email, password });
+    },
+
+    onSuccess: (data) => {
+      if (data) {
+        localStorage.setItem("token", data.data.token.access_token);
+        localStorage.setItem("refreshToken", data.data.token.refresh_token);
+      }
+    },
+    onError: (error) => {
+      if (axios.isAxiosError(error)) {
+        toast.error(error.response?.data.message || "Something went wrong");
+      }
     },
   });
 }
